tokengen: allow to configure the number of threads
Add an options panel, TokenOptionsPanel, to allow to configure the
number of threads used during the token generation.
Change ExtensionTokenGen to hook the TokenParam and TokenOptionsPanel.
Change TokenParam to:
 - Extend VersionedAbstractParam, so that it can be versioned;
 - Rename the property/method from threadPerScan to threadsPerScan;
 - Add constant for the default number of threads and change it from 20
 to 5 (same as other tools, e.g. fuzzer);

Add a button to TokenPanel to allow direct access to the options.
Add help page for the options and update main page to link to it.
Update changes in ZapAddOn.xml file.