fuzz: correct payload count for random regex
Change RegexPayloadGenerator to take into account the random state when
calculating the number of payloads that are generated (with random it
generates into the limit specified).
Change RegexPayloadGeneratorUIHandler to prevent setting the max/limit
to 0 when random, as that prevents the generation of any payload. Also,
take into account the random state when calculating the maximum number
of payloads that should be persisted and when previewing the payloads.