accessControl: respect the mode (#1538)
Change ExtensionAccessControl to stop all the scans when the mode is
changed to safe or if the scans are not in scope if changed to
protected. Also, prevent starting the scans if the mode does not allow
it and notify the panel of mode changes.
Change AccessControlScanOptionsDialog to warn (and prevent the scan) if
the mode does not allow to start it.
Update help to mention the mode restrictions.
Update ZapAddOn.xml file.