websocket: don't wait forever for persisted msg (#1550)
Change ExtensionWebSocket to not wait forever for the persisted hanshake
message, if the message is not persisted the proxy thread would hang
forever. It will now wait at most 1 second, which should be enough time
(if not it simply doesn't show the handshake message). Also, core is now
persisting the message synchronously (zaproxy/zaproxy#4586) which makes
the wait obsolete.