soap: dynamically unload the add-on (#1582)
Change ExtensionImportWSDL to remove the spider parser when unloaded and
declare that it can be unloaded (it was the last change need to
dynamically unload the add-on). Also, remove unnecessary log when adding
the parser, the presence of the spider already indicates if the parser
is added.
Remove code no longer needed in the spider parser and corresponding test
case, as the parser is removed it not longer needs to be enabled and
disabled.
Update changes in ZapAddOn.xml file.