soap: fix and tidy up tests
Fix failing tests in ImportWSDLTestCase (by not depending on the order
of the tests) and ExtensionImportWSDLTestCase (by loading the resource
messages).
Change WSDLCustomParserTestCase to not use reflection nor depend on the
order of the tests.
Rename SOAPActionSpoofingActiveScannerTest to have the same suffix as
the other tests ("TestCase", to use the same Ant include rule), also,
change it to not use reflection.
Tweak WSDLCustomParser and SOAPActionSpoofingActiveScanner to ease the
tests, by relaxing accessibility of the methods or return proper values.
Change build.xml file to run the soap tests (and copy its resources).