openapi: use UTF-8 and ignore BOM
Change ExtensionOpenApi to read the file with UTF-8 instead of relying
on the default charset.
Change SwaggerConverter to remove the BOM from the definition, the
Swagger library checks the first character to decide if the contents
should be parsed as JSON or YAML.
Update changes in ZapAddOn.xml file.

Fix zaproxy/zaproxy#4676 - openAPI import error