groovy: rename add-on for broader usage
Change the add-on name (and description) to Groovy Support to not focus
just on scripting, the add-on is also useful as a dependency for add-ons
written in Groovy.
Update help to mention that use case and reflect the name/description
changes.
Change version to follow semantic versioning, to make it reliable for
other add-ons to depend on.
Update changes in ZapAddOn.xml file.