scripts: tweak help
Remove recommendation to copy JARs to lib directory, that will not work
by default with newer ZAP versions (the classpath needs to be changed to
include them).
Change the TOC, index, headers, and titles to use Script Console instead
of Scripts, so they match the add-on name (as with other add-ons).

Related to zaproxy/zaproxy#4755 - Stop relying on ClassLoaderUtil