customreport: show dialogue, always
Change OptionDialog to always (on cancel, accept, and close) clear its
own reference in ExtensionCustomReport to ensure that it's always shown
when invoked through the menu, otherwise if closed through the window
decorations it will not be shown anymore.
Change ExtensionCustomReport to not expose the method that clears the
reference to the dialogue as that's handled internally also create the
dialogue when invoked through the menu (instead of using a public
method).