Update Nano-Httpd to version 2.3.1
Update Nanohttpd from Version 2.1.1 to 2.3.1. There are two major changes in new nanohttpd version. The first one change is in NanoHTTPD.Response class. The new version offers a static method (newFixedLenghtResponse) in order to create a fixed length response. The second one is getParms method which is deprecated. The method replaced by getParameters which returns a Map<String,List<String>>. In order to address the last one change, a new method (getFirstParamValue) was added to NanoServerHandler because most of test cases use only the first value of parameters.