Issue 4924 - CacheableScanner Fix NumberFormatException
CacheableScanner - When parsing "lifetime" (max-age) value split on
comma and use the 0th item in case there's a malformed value passed.
(This works whether or not there's a comma to split on. The string was
already split on space, and has been checked to start with max-age.) If
a NFE still occurs (exceed Long.MAX_VALUE or still non-number
characters) mark lifetime as not found and log a debug message.
ZapAddOn.xml - Updated change section.

Fixes zaproxy/zaproxy#4924