Remove unreachable strength/threshold cases
Remove switch cases using `DEFAULT` or `OFF` for `AttackStrength` and
`AlertThreshold`, those cases do not happen, the methods used to obtain
the strength/threshold return the value configured for the default
strength (for example, MEDIUM) and the `OFF` is used to disable the
scanners so the scanner would not even be called in that case.
Update changes in ZapAddOn.xml file (where required).