ascanrules: correct time replacement
Change CommandInjectionPlugin to replace the time token directly without
using MessageFormat, some of the payloads are quoted which would prevent
the replacement (while one could escape the quotes and keep using the
MessageFormat it would make the payloads harder to read and maintain).
Add test to assert the expected behaviour.
Update changes in ZapAddOn.xml file.