selenium: disable JSON viewer for AJAX Spider
Disable Firefox JSON viewer when returning a WebDriver for AJAX Spider
to prevent the spider from crawling it.
Add static methods to ExtensionSelenium to allow to get a WebDriver with
a requester and Browser, to be used by BuiltInSingleWebDriverProvider.
Change BuiltInSingleWebDriverProvider to call the new methods.
Update changes in ZapAddOn.xml file.