quickstart: inform when disabled by current mode
Change QuickStartPanel to show a message in the URL field when the mode
does not allow to start the attack, also, disable the button that allows
to select a target.
Change ExtensionQuickStart to not initialise the view components if
there's no view when the mode is changed.
Update changes in ZapAddOn.xml file.

Fix zaproxy/zaproxy#5069 - Add a message to explain why Quick Scan not
available in protected etc modes