Update API generator for 2.7.0 and 2.8.0
Move ApiGenerator to api package to access core class.
Add Websocket API to ApiGenerator.
Change ApiGenerator to make use of the new core method (>=2.8.0) which
allows to pass a ResourceBundle, to include the descriptions of the API
endpoints being generated.
Add dummy Messages.properties to be able to work with ZAP 2.7.0.
Update generate-apis in build.xml file to include just the required
files on the classpath and to depend on compile to work by itself.

For zaproxy/zaproxy#5044 - NodeJS API Upgrade