zest: rely on script context writer
Change ZestZapRunner to not set the script console writer as output
writer if the script context writer is already set, while both end up in
the script console the latter allows the core to intercept the writes.
For the cases the context writer is not set the writes are not
intercepted.
Update changes in ZapAddOn.xml file.

Related to zaproxy/zaproxy#5113 - Allow to differentiate scripts' output