Add default assertions to passive alerts (beta)
Change InsecureAuthenticationScan to not set the attack field in the
alert, instead mention that in the description.
Change CharsetMismatchScanner to use just one resource message for the
description, it's clearer.
Change PassiveScannerTestUtils to assert that the alerts raised have the
attack field empty and the expected scanner ID.

Related to zaproxy/zaproxy#5061 and referenced issues.