Add default assertions to passive alerts (alpha)
Change UserControlledCharsetScanner to not set the attack field in the
alert, instead mention that in the description. Also, remove unused
parameter.
Change UserControlledCookieScanner to not set the attack field in the
alert, instead mention that in the extra info. Fix zaproxy/zaproxy#5149.
Change InsecureFormLoadScanner, UserControlledHTMLAttributesScanner,
InsecureFormPostScanner, and UserControlledJavascriptEventScanner to not
read the (empty) resource message for the attack field.
Remove other empty and unused attack resource messages, should not be
needed.
Change PassiveScannerTestUtils to assert that the alerts raised have the
attack field empty and the expected scanner ID.

Related to zaproxy/zaproxy#5061 and referenced issues.