ascanrulesBeta: fix missing resource messages
Change CrossDomainScanner to return empty strings for the description
and solution instead of trying to get the resource messages (which do
not exist), the alerts being raised do not use those methods but
external code might still call them which would lead to exception or
errors.
Update changes in ZapAddOn.xml file.