jxbrowser: tweak messages and log browser version
Fix typo in the initial message shown in the browser window (missing
space between "browsers" and "to").
Remove references to the ChromeDriver and Selenium help page in the
message that indicates that the browser failed to start, the driver is
now bundled in the JxBrowser add-ons.
Log the JxBrowser version used by the add-on, to be easier to know which
version is in use.