Move common test code to a project
Add a top level project, `testutils`, that provides the common test
utilities for the add-ons. The add-ons will depend on this project as
needed.
Remove all test libraries (`testlib` dir), they will be obtained from
Maven repository.
Correct the path to the scripts in `AbstractVerifyScriptTemplates` to
match the new directory layout.