fuzzdb: relocate to its own project
Move the files to its own project, `fuzzdb` under `addOns`.
Remove the add-on manifest (`ZapAddOn.xml`) and `FuzzdbFiles` (helper
class to create the list of home files), the manifest will be generated
by the build.
Add build file (`fuzzdb.gradle.kts`) to the project to configure the
add-on.
Add `CHANGELOG.md` to the project (generated from ZAP marketplace).