websocket: relocate to its own project
Move the files to its own project, `websocket` under `addOns`.
Remove the add-on manifest (`ZapAddOn.xml`), it will be generated by the
build.
Update repository path (per relocation) in about help page.
Add build file (`websocket.gradle.kts`) to the project to configure the
add-on.
Add `CHANGELOG.md` to the project (generated from ZAP marketplace).
Add the add-on to the list of main add-ons in `addOns.gradle.kts`.