ascanrulesBeta: relocate to its own project
Move the files to its own project, `ascanrulesBeta` under `addOns`.
Remove the Procyon library (add-on `lib` directory), it will be obtained
from Maven repository (the library without command line support, which
is not needed for the current usage).
Remove the add-on manifest (`ZapAddOn.xml`), it will be generated by the
build.
Add build file (`ascanrulesBeta.gradle.kts`) to the project to configure
the add-on.
Add `CHANGELOG.md` to the project (generated from ZAP marketplace).
Add the add-on to the list of weekly add-ons in `addOns.gradle.kts`.