beanshell: relocate to its own project
Move the files to its own project, `beanshell` under `addOns`.
Remove the Beanshell library (add-on `lib` directory), it will be
obtained from Maven repository.
Remove the add-on manifest (`ZapAddOn.xml`), it will be generated by the
build.
Add build file (`beanshell.gradle.kts`) to the project to configure the
add-on.
Add `CHANGELOG.md` to the project (generated from ZAP marketplace).