fuzz: relocate to its own project
Move the files to its own project, `fuzz` under `addOns`.
Remove all libraries (add-on `lib` directory) except (forked) Generex,
they will be obtained from Maven repository. The Generex library was
moved to a project top level directory `lib`.
Replace `jbrofuzz` with `jbrofuzz-core` (and do not include transitive
dependencies), only the core library is needed.
Remove the add-on manifest (`ZapAddOn.xml`), it will be generated by the
build.
Add build file (`fuzz.gradle.kts`) to the project to configure the
add-on.
Add `CHANGELOG.md` to the project (generated from ZAP marketplace).
Add the add-on to the list of main add-ons in `addOns.gradle.kts`.