jython: relocate to its own project
Move the files to its own project, `jython` under `addOns`.
Remove the Jython library (add-on `lib` directory), it will be obtained
from Maven repository.
Remove the add-on manifest (`ZapAddOn.xml`), it will be generated by the
build.
Add build file (`jython.gradle.kts`) to the project to configure the
add-on.
Add `CHANGELOG.md` to the project (generated from ZAP marketplace).