ascanrulesAlpha: relocate to its own project
Move the files to its own project, `ascanrulesAlpha` under `addOns`.
Remove the jsoup library (add-on `lib` directory), it will be obtained
from Maven repository.
Remove the add-on manifest (`ZapAddOn.xml`), it will be generated by the
build.
Add build file (`ascanrulesAlpha.gradle.kts`) to the project to
configure the add-on.
Add `CHANGELOG.md` to the project (generated from ZAP marketplace).