jsonview: relocate to its own project
Move the files to its own project, `jsonview` under `addOns`.
Move the `README.md` to the top level directory of the project.
Remove the add-on manifest (`ZapAddOn.xml`), it will be generated by the
build.
Add build file (`jsonview.gradle.kts`) to the project to configure the
add-on.
Add `CHANGELOG.md` to the project (generated from ZAP marketplace).