jxbrowserlinux64: relocate to its own project
Move the files to its own project, `jxbrowserlinux64` under
`addOns:jxbrowsers`.
Remove the add-on manifest (`ZapAddOn.xml`), it will be generated by the
build.
Add build file (`jxbrowserlinux64.gradle.kts`) to the project to
configure the add-on.
Add `CHANGELOG.md` to the project (generated from ZAP marketplace).
Add the add-on to the list of main add-ons in `addOns.gradle.kts`.