Downgrade source code disclosure to Medium
In my view disclosing the source code is not as serious as things like
XSS and SQLi. It may well be a serious problem for the company involved
but it does not mean that the attacker will definitely be able to find
other vulnerabilities, especially if the code is already open source.

Happy to discuss this further if anyone disagrees...

Signed-off-by: Simon Bennetts <psiinon@gmail.com>