Create unit test for InformationDisclosureSuspiciousComments beta (#2084)
Create unit test for InformationDisclosureSuspiciousComments beta
scanner. Enhance this scanner to ignore whitespace before and after any
comment words in the suspicious commments file, and a few other minor
tweaks. Add a comment to the suspicious comments file to indicate that
whitespace before or after any values in that file will be ignored.

Close zaproxy/zaproxy#2805.