customreport: fix dialogue layout
Add the tabbed panel to the centre of the dialogue so that its tabs are
properly (re)sized.
Tweak the Scope tab to let the fields expand/shrink when the dialogue is
resized, also, give extra vertical space to the description field (the
biggest).

Fix zaproxy/zaproxy#5521 - Customizable HTML report is not working

Signed-off-by: thc202 <thc202@gmail.com>