Fix IDE error on package name collision
TestCrossSiteScriptV2UnitTest was colliding with the resource package
(path) by the same name. Renaming it to lowercase successfully addresses
the collision.

I also looked in to removing the "UnitTest" portion of the path name
since it's in a obvious 'Test' section anyway, however that causes other
failures when the classnames mismatch and resource access fails.

Signed-off-by: kingthorin <kingthorin@users.noreply.github.com>