Create Common Library add-on
Move the project `sharedutils` to an add-on to allow any ZAP add-on to
use/depend on it and to avoid duplicating the classes/resources/libs in
all of them.
Make classes that are not intended to be extended final and make some
fields private final.
Change `BinList` to a singleton and let it be accessed directly.
Add minimal help.
Remove `SetCookieUtils`, a subset of `CookieUtils`.
Add JavaDoc to some classes/methods and add since JavaDoc tag.

The add-on will be published to Maven Central with the following
group/id:
 - Group: `org.zaproxy.addon`;
 - Id: `commonlib` (same as the add-on ID).

Update add-ons that were already using the `sharedutils` project to
use/depend on the add-on.

Signed-off-by: thc202 <thc202@gmail.com>