Suspicious Comments rule alert tweaks + release
The last set of changes to this rule exposed a core bug:
zaproxy/zaproxy#6421

Examining the alerts raised in that case showed that the info included
was actually not that useful.
An issue would be raise for every instance of a suspicious string in a
page, and the evidence was the relevant page element, which for scripts
could be very large. And scripts could trigger the bug above ;)

This PR changes the rule to only raise one alert per suspicious string
instance per page - if there are more than one instance then it gives
the count. The evidence is just the string matched while the 'other'
field now contains the first 128 bytes of the first relevant element.
Screen shots to be included...

Also prepare add-on for release so that we're less likely to hit the
core bug...

Signed-off-by: Simon Bennetts <psiinon@gmail.com>