Reports - modern template WIP
Based on the templates in #2876 but implemented as themes.
The plan is to replace the tradition-html-plus template with this one.

I havnt done all of the themes yet, and they arent yet _exactly_ the
same as the originals.
The chart is also much bigger - not sure if thats better or worse though
;)

Feedback on the approach appreciated...

Signed-off-by: Simon Bennetts <psiinon@gmail.com>