Reports: Add traditional JSON report
The JSON reports should be equivalent to the existing core ones with the
exception of the formating. The escaping should be exactly the same.
The aim is to ensure anyone who relies on these reports should not be
impacted when we switch over.

I'll need to make similar changes to the XML report.

Signed-off-by: Simon Bennetts <psiinon@gmail.com>