Reports: OutputSummary config file support phase 1
Initial support for the packaged scan config files.
Supports setting the alerts to IGNORE/INFO/WARN/FAIL.
It does not yet support the OUTOFSCOPE elements so will not be enabled
in the baseline scan just yet.

However this PR also includes some tweaks to make the output more
backwards compatible, based on tests created for the baseline scan.

Signed-off-by: Simon Bennetts <psiinon@gmail.com>