fix: domxss: Check alert dialog content to reduce false positives
- DomXssScanRule > Now gets the alert dialog's content before catching
UnhandledAlertException then compares it to ensure the dialog is likely
to be one ZAP caused. Increased confidence to HIGH.
- DomXswsScanRuleUnitTest > Added a single method to test an unrelated
alert dialog condition. (Ex: When an alert shouldn't pop.) Adjusted to
accommodate new HIGH confidence. Corrected one documentation URL.
- DomAlertException > Renamed to DomAlertInfo, and changed control away
from Throwing a custom exception to checking if the return is null or
not.
- ExtensionSelenium > Added a utility method to accept a capabilities
consumer to be passed. In the case of the domxss add-on this
sets the unhandled dialog behavior to IGNORE (vs the seeming default of
ACCEPT).
- RandomAlert.html > Supports the added unit test.
- CHANGELOG.md > Added fix note.

Signed-off-by: kingthorin <kingthorin@users.noreply.github.com>