openapi: import from file even if it has issues
Change `SwaggerConverter` to return as soon as it's able to parse the
definition, even if it has issues.
Change `ExtensionOpenApi` to not skip the parsed file definition when it
has issues, still try to import what was parsed.
Restore behaviour of previous versions and match the behaviour when
importing from URL.

Fix zaproxy/zaproxy#6758.

Signed-off-by: thc202 <thc202@gmail.com>