improve: soap: Various User friendliness changes
- WSDL File scan rule now includes identification of URLs in the
form "example.com/service?wsdl", and specifically excludes 404s and 500s
(including support for Custom Pages).
- Import a WSDL file from a URL now logs a more specific message for
some exception conditions, or when failing to convert the
input 'URL' string into an actual URL to make the WSDL request.
- The import from URL dialog now accepts pressing the enter key to
activate the import process (instead of having to click the Import
button), and the escape key to close/cancel the dialog.
- Changed Scan Rule test class to extend PassiveScannerTest, adjusted
some test names to be more specific, added tests to assert the new
behavior.

Signed-off-by: kingthorin <kingthorin@users.noreply.github.com>