ascanrules: XSS: try more payloads if outside of HTML tags
In the XSS addOn, in the case when the payload is reflected outside…f
HTML tags, try more payloads.

I made this to catch a potential XSS that I noticed being missed by the
active scan in an API that I was scanning. Payload was reflected in a
json response.

Signed-off-by: wojtek <dziewiecki@code-intelligence.com>