network: correct handlers result
Return immediately if a handler overrides the message.
Do not override when setting an error response nor if legacy listeners
set a response to allow other handlers to be notified.
Do not send a message if it already has a response.
Be more clear of the outcome of override and close in the JavaDoc.

Signed-off-by: thc202 <thc202@gmail.com>