network: fix issues
Use `localhost` when migrating core proxy options if the address is not
present, instead of "any address".
Use one of the local addresses when providing the main proxy address to
other extensions, when the main proxy is using "any address".
Set the core proxy address/port when changes are done to the main
proxy.

Signed-off-by: thc202 <thc202@gmail.com>