Quit all browsers when ZAP shuts down
Change `ExtensionSelenium` to quit all WebDrivers not just the ones
that were created to proxy through the main proxy, also, quit them
later when the extension is destroyed, to allow other extensions to
stop gracefully.
Change AJAX Spider to stop the spiders (GUI and API) on stop, so they
are stopped always previously just stopped when the add-on was
uninstalled.
Change `ExtensionDomXSS` to stop the server on stop not just when the
add-on is uninstalled.

Fix zaproxy/zaproxy#6643.

Signed-off-by: thc202 <thc202@gmail.com>