oast: Fix flaky unit test
The issue was happening because of a race condition between the test and
the poller which was running in a different thread. They both were
calling the `getInteractions` method which led to the statistic being
incremented twice. When the poller called the method before the test,
the assertion failed because it was only expecting a single increment
but found two.

The fix was to disable the polling thread for the test.

Signed-off-by: ricekot <github@ricekot.com>