Ascanrules: Fix XSS false positive in textarea
A script tag will not run inside a textarea, previously the code would
incorrectly report it as an XSS.

As part of this change I've copied HtmlContext and HtmlContextAnalyser
from the core - I'll deprecate them in another PR.

Tested against Firing Range (no change) and WebSecLab locally - this
fixes 2 of the FP's ZAP currently reports.
The plan is to publish the WebSecLab results on the website.

Signed-off-by: Simon Bennetts <psiinon@gmail.com>