spider: provide context to parsers
Provide all the information needed (context) to the parsers using an
object to allow to provide more data as needed without breaking binary
compatibility.
Update all parsers to use the context to get the information needed.
Rely on the context for default values when reporting URLs found.
Remove adapter for compatibility with 2.11 since the add-on is no
longer targeted for that release.

Signed-off-by: thc202 <thc202@gmail.com>