requester: use own classes for (re)send dialogues
Add classes for the send and resend dialogues, which supersede the core
ones.
Load the send dialogue and add the history reference info when the core
extension is deprecated.
Remove unnecessary methods and the interface `MessageSender`, let the
implementations handle how the messages are sent, it was also exposing
implementation details (e.g. clean up was only needed for HTTP and it
will soon no longer be needed per newer HTTP client).
Rename some methods to be message agnostic and others to be more clear
what they do.

Signed-off-by: thc202 <thc202@gmail.com>