requester: relocate classes and resources
Move the classes to be under the newer package followed by other
add-ons, also, move classes not intended to be used by other add-ons to
be under an internal package.
Rename a class to be generic ("message" instead of "request").
Move resources to match the location of the classes.
Update the changelog with the functionality provided and mention the
add-on is now following semantic versioning.
Change the description of the extension and add-on to better match the
provided functionality.

Signed-off-by: thc202 <thc202@gmail.com>