reports: add "XML with requests and responses" report template (#3983)
For JSON and HTML reports, a "plus" version of the template exists that
also includes the requests and responses.  It would be useful to have a
similar feature for XML reports, thus, I created one here, and will also
be adding support for this format to DefectDojo (if it gets merged
here).

Signed-off-by: Max Maass <max.maass@iteratec.com>
Signed-off-by: tuxBurner <s.hardt@micromata.de>