ascanrules: address XSS false negative
Only require that the attribute value starts with the javascript scheme
injection (ignoring leading whitespaces), the presence of other
characters following the injection do not necessarily invalidate it.
For example, if it's also injected with a comment (e.g. `//`) it would
make it valid, even if it has more/other characters after.

Signed-off-by: thc202 <thc202@gmail.com>