postman: Fix class loader issue
All add-ons have their own class loader and they are isolated from each
other (unless they explicitly depend on others), with the move of
Jackson libraries to commonlib add-on Jackson started to use the class
loader of commonlib instead of postman which caused the deserialisation
to fail to find the postman classes (as commonlib does not depend on
postman). The change done in PostmanParser ensures Jackson uses the
postman class loader and thus able to find the postman classes. The
change in ListDeserializer is just to use the same configuration we set
up initially (otherwise we would have to configure the ObjectMapper once
again).

Thanks @thc202 for the fix and explanation :)

Signed-off-by: VitikaSoni <vitika.program@gmail.com>