network: ensure servers stopped before starting
Stop the servers before attempting to start them, the servers are
started after the installation of the add-on but if done during startup
core will execute the command line arguments of the extension which
would start the servers again making the servers fail to bind (e.g.
address already in use) and ZAP fail to start.

Signed-off-by: thc202 <thc202@gmail.com>