Move `ValueGenerator` from core and rename it
Move the `ValueGenerator` interface to the `commonlib` add-on and
rename it to `ValueProvider` to avoid duplicated names. Add the default
implementation as well.
Change all other add-ons to use the interface/implementation from the
`commonlib` add-on.
Change the `formhandler` add-on to depend on `commonlib` and set the
implementation that provides the custom values.

Part of zaproxy/zaproxy#8016.

Signed-off-by: thc202 <thc202@gmail.com>