client: shutdown spider thread pool
Ensure that no non-daemon threads are kept running after the spider
finishes, which would lead to thread leaks and in command line mode
could prevent ZAP from terminate gracefully.
Ensure that the WebDriver is returned and the task completed even in
case of exceptions, otherwise that would prevent the spider from
finishing.

Signed-off-by: thc202 <thc202@gmail.com>