client: address concurrency issues with tasks
Sync the access internally in the task model to, regardless of how ZAP
is run, prevent concurrency issues which would lead to some tasks going
missing while they were being added to the list of tasks which would
later lead to exceptions as the tasks were not present in the model.
Only report model changes if there's a view.

Signed-off-by: thc202 <thc202@gmail.com>