client: adjust thread accesses
Reduce the work done in the EDT as much of it does not need to happen
there.
Sync addition of the tasks as well to ensure the check done on removal
takes into account the tasks that might have been being added/executed
at the same time.
Also, remove sync on WebDriver retrieval since that's already synced
internally.
Rework the task model to make the ID/row match in all cases, it was
still possible for the ID/row to be interleaved enough that later would
lead to index out of bound exceptions.

Signed-off-by: thc202 <thc202@gmail.com>